fix(solver): split augmented index redirect helper
